requirements.txt
# Web server
fastapi==0.129.*
uvicorn==0.40.*
httpx==0.28.*

# Supabase
supabase==2.24.*

# LLM + Embedding (Gemini API)
google-genai==1.*

# Data collection
dropbox==12.*
imap-tools==1.*

# Text extraction
PyMuPDF==1.27.*
python-docx==1.*
openpyxl==3.*
python-pptx==1.*
beautifulsoup4==4.*
gethwp==1.*
pyhwp==0.1b15
lxml

# Utilities
python-dotenv==1.*
pydantic==2.*
pydantic-settings
